started outsourcing event react hook
