--- packages/event/src/react/index.ts
@@ -1,1 +1,0 @@
-export { useEvent } from './hooks/useEvent';
