+++ packages/react/src/core/hocs/AgileHOC.ts
@@ -10,7 +10,7 @@ import Agile, {
   normalizeArray,
 } from '@agile-ts/core';
 import type { Collection } from '@agile-ts/core'; // Only import Collection and Group type for better Treeshaking
-import { LogCodeManager } from '../logCodeManager';
+import { LogCodeManager } from '../../logCodeManager';
 
 /**
  * A Higher order Component for binding the most relevant value of multiple Agile Instances
