+++ packages/react/src/index.ts
@@ -1,10 +1,5 @@
 import reactIntegration from './react.integration';
 
-export { useAgile } from './hooks/useAgile';
-export { AgileHOC } from './hocs/AgileHOC';
-export { useWatcher } from './hooks/useWatcher';
-export { useProxy } from './hooks/useProxy';
-export { useValue } from './hooks/useValue';
-export { useSelector } from './hooks/useSelector';
+export * from './core';
 
 export default reactIntegration;
