+++ packages/react/src/proxy/index.ts
@@ -1,0 +1,1 @@
+export * from './hooks/useProxy';
